huge images: imageio and the like (cherry picked from commit 72b11b23e852da394e10f64c639c07dcd4e39776)

+++ src/common/imageio.h
@@ -74,7 +74,7 @@ dt_imageio_export_with_flags(
   const char                        *filter,
   const gboolean                     copy_metadata);
 
-int dt_imageio_write_pos(int i, int j, int wd, int ht, float fwd, float fht, int orientation);
+size_t dt_imageio_write_pos(int i, int j, int wd, int ht, float fwd, float fht, int orientation);
 
 // general, efficient buffer flipping function using memcopies
 void
